refactor : adminCotroller 로 통합
