+++ src/main/java/kr/co/iei/admin/controller/AdminController.java
@@ -1,4 +1,4 @@
-package kr.co.iei.event.cotroller;
+package kr.co.iei.admin.controller;
 
 import java.io.File;
 import java.util.List;
@@ -19,38 +19,42 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
-import kr.co.iei.event.model.dto.EventDTO;
-import kr.co.iei.event.model.service.EventService;
+import kr.co.iei.admin.model.dto.EventDTO;
+import kr.co.iei.admin.model.service.EventService;
+import kr.co.iei.review.model.dto.ReviewDTO;
+import kr.co.iei.review.model.service.ReviewService;
 import kr.co.iei.util.FileUtils;
 
 @CrossOrigin("*")
 @RestController
-@RequestMapping(value="/event")
-public class EventController {
+@RequestMapping(value="/admin")
+public class AdminController {
 	@Autowired
 	private FileUtils fileUtils;
 	@Autowired
 	private EventService eventService;
+	@Autowired
+	private ReviewService reviewService;
 	@Value("${file.root}")
 	private String root;
-	@GetMapping("/{month}")
+	@GetMapping("/event/{month}")
 	public ResponseEntity<List> selectMonthEvent(@PathVariable String month){
 		List eventList = eventService.selectMonthEvent(month);
 		return ResponseEntity.ok(eventList);
 	}
-	@GetMapping("/onGoing")
+	@GetMapping("/event/onGoing")
 	public ResponseEntity<List> selectOnGoingEvent(@RequestParam String date){
 		List eventList = eventService.selectOnGoingEvent(date);
 		return ResponseEntity.ok(eventList);
 	}
-	@GetMapping("/end")
+	@GetMapping("/event/end")
 	public ResponseEntity<List> selectEndEvent(@RequestParam String date){
 		System.out.println(date);
 		List eventList = eventService.selectEndEvent(date);
 		return ResponseEntity.ok(eventList);
 	}
 	
-	@PostMapping
+	@PostMapping("/event")
 	public ResponseEntity<Integer> insertEvent(@ModelAttribute EventDTO event,@ModelAttribute MultipartFile img){
 		String savepath = root +"/event/thumb/";
 		String filepath = fileUtils.upload(savepath, img);
@@ -58,7 +62,7 @@ public class EventController {
 		int result = eventService.insertEvent(event);
 		return ResponseEntity.ok(result);
 	}
-	@PatchMapping("/{eventNo}")
+	@PatchMapping("/event/{eventNo}")
 	public ResponseEntity<Integer> updateEvent(@ModelAttribute EventDTO event,@ModelAttribute MultipartFile img,@PathVariable int eventNo){
 		if(img!=null) {
 			String savepath = root +"/event/thumb/";
@@ -74,7 +78,7 @@ public class EventController {
 		}
 		return ResponseEntity.ok(1);
 	}
-	@DeleteMapping("/{eventNo}")
+	@DeleteMapping("/event/{eventNo}")
 	public ResponseEntity<Integer> deleteEvent(@PathVariable int eventNo){
 		String filepath = eventService.deleteEvent(eventNo);
 		File file = new File(root+"/event/thumb/"+filepath);
@@ -82,5 +86,10 @@ public class EventController {
 			file.delete();
 		}
 		return ResponseEntity.ok(1);
+	}
+	@GetMapping("/report")
+	public ResponseEntity<List> reportedReview(){
+		List reportedReviews = reviewService.selectReportedReview();
+		return ResponseEntity.ok(reportedReviews);
 	}
 }
