+++ src/main/resources/mapper/event-mapper.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="kr.co.iei.event.model.dao.EventDao">
+<mapper namespace="kr.co.iei.admin.model.dao.EventDao">
 	<select id="selectOnGoingEvent" resultType="event">
 		select * from event 
 		where to_date(#{date}, 'YYYY-MM-DD')
